Restructured the code, added a search box to the codes group editor
